Got USART2 interrupt working...
Quite the wild goose chase... in the end I think using OpenOCD that's
NOT part of the OpenSTM32.org toolchain was causing me debugging grief.
Switched the Debugging configuration to "Ac6 STM32 Debugging" was
what allowed setting breakpoints inside HAL_UART_RxCpltCallback().

Either that or initiating ISR base reading with HAL_UART_Receive_IT()
it the hot ticket.  Not sure yet.  Anyway, it's working now!

Will try going back to previous debugging configuration to see if
was broken or not.