Add double buffering to gyro in preparation for DMA use.
DMA *may* be overkill, but maybe not.  Certainly ISR based
reading of I2C is a good thing, for high data rate / low latency,
but DMA may not provide enough extra CPU cycles to be worth the effort.
Only way to know is to make it happen and compare the results.