Added USARTTask...
USARTTask waits on a FreeRTOS queue, which contains one
iteration of sensor data (minus the barometer) from the
Adafruit 1604 IMU unit. The ImuRead() function copies the most
recently read data from the sensors to a tmp buff that it sends
to the queue. This copying adds approx. 112us to the sensor
read loop, putting it slightly over 1ms.  But that's at 80Mhz
on the STM32L4 processor.  STM32F4 will have no problem at 168Mhz.
This could be optimized.  If it turns out that I care.